# backend/docs/swagger.yaml
basePath: /v1
definitions:
  analogdb.Camera:
    properties:
      created:
        example: "2025-07-11T12:00:00Z"
        type: string
      description:
        example: Nikon FM2 is a dependable SLR camera
        type: string
      id:
        example: 1
        type: integer
      make:
        example: nikon
        type: string
      model:
        example: fm2
        type: string
      post_count:
        example: 50
        type: integer
      slug:
        example: nikon-fm2
        type: string
      top_posts:
        items:
          $ref: '#/definitions/analogdb.CatalogPost'
        type: array
      updated:
        example: "2025-07-11T12:00:00Z"
        type: string
    type: object
  analogdb.CatalogPost:
    properties:
      id:
        example: 1
        type: integer
      images:
        items:
          $ref: '#/definitions/analogdb.Image'
        type: array
      score:
        example: 150
        type: integer
      title:
        example: Sunset over the lake
        type: string
    type: object
...
  analogdb.CreateCamera:
    properties:
      description:
        example: Nikon FM2 is a dependable SLR camera
        type: string
      make:
        example: nikon
        type: string
      model:
        example: fm2
        type: string
    type: object
...
  analogdb.Error:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: not found
        type: string
    type: object
...
  analogdb.Image:
    properties:
      height:
        example: 7600
        type: integer
      resolution:
        example: high
        type: string
      url:
        example: https://cloudfront.net/1.jpeg
        type: string
      width:
        example: 4800
        type: integer
    type: object
...
  analogdb.PostExtraction:
    properties:
      created:
        type: string
      extractor_version:
        example: 2026-10-a
        type: string
      input:
        example: 'title: Dusk [Nikon FM, Portra 400]'
        type: string
      input_hash:
        example: 9f2c...
        type: string
      model:
        example: google/gemini-2.5-flash-lite
        type: string
      post_id:
        example: 1234
        type: integer
      raw:
        type: object
      unmatched:
        items:
          type: object
        type: array
      updated:
        type: string
    type: object
  server.CamerasResponse:
    properties:
      cameras:
        items:
          $ref: '#/definitions/analogdb.Camera'
        type: array
    type: object
...
  server.CreateCameraResponse:
    properties:
      camera:
        $ref: '#/definitions/analogdb.CreateCamera'
      message:
        example: Success, camera created
        type: string
    type: object
...
  server.DeleteResponse:
    properties:
      message:
        example: Success, post deleted
        type: string
    type: object
  server.ExtractionsRequest:
    properties:
      extractions:
        items:
          $ref: '#/definitions/analogdb.PostExtraction'
        type: array
    type: object
  server.ExtractionsResponse:
    properties:
      extractions:
        items:
          $ref: '#/definitions/analogdb.PostExtraction'
        type: array
      next_before_id:
        example: 1200
        type: integer
    type: object
...
  server.UpsertExtractionsResponse:
    properties:
      skipped:
        example:
        - 1234
        items:
          type: integer
        type: array
      written:
        example: 499
        type: integer
    type: object
...
host: api.analogdb.com
info:
  contact: {}
  description: API for analogdb film database
  title: AnalogDB API
  version: "1.0"
paths:
  /admin/extractions:
    get:
      description: List stored metadata extractions, newest post first (requires authentication)
      parameters:
      - description: Only extractions with or without unmatched mentions
        in: query
        name: has_unmatched
        type: boolean
      - description: Only extractions with an unmatched mention of this kind
        enum:
        - camera
        - film
        in: query
        name: kind
        type: string
      - description: Only extractions with an unmatched mention with this normalized
          key
        in: query
        name: key
        type: string
      - description: Include the input text and raw LLM output
        in: query
        name: full
        type: boolean
      - description: Only post ids below this, for paging
        in: query
        name: before_id
        type: integer
      - default: 100
        description: Number of results, at most 500
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ExtractionsResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: List post metadata extractions
      tags:
      - extractions
    post:
      consumes:
      - application/json
      description: Create or replace the stored metadata extraction of each post (requires
        authentication). Posts that don't exist are skipped.
      parameters:
      - description: extractions to store, at most 500
        in: body
        name: extractions
        required: true
        schema:
          $ref: '#/definitions/server.ExtractionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.UpsertExtractionsResponse'
        "400":
          description: Invalid extractions
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Store post metadata extractions
      tags:
      - extractions
  /camera:
    post:
      consumes:
      - application/json
      description: Create a new camera entry (requires authentication)
      parameters:
      - description: camera data to create
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/analogdb.CreateCamera'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreateCameraResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Create a new camera
      tags:
      - camera
    put:
      consumes:
      - application/json
      description: Create a new camera entry (requires authentication)
      parameters:
      - description: camera data to create
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/analogdb.CreateCamera'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreateCameraResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Create a new camera
      tags:
      - camera
  /camera/{id}:
    delete:
      description: Delete a catalog camera by ID (requires authentication). Refused
        while any post still uses it.
      parameters:
      - description: Camera ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DeleteResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/analogdb.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/analogdb.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/analogdb.Error'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/analogdb.Error'
        "409":
          description: Still used by posts
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      security:
      - BasicAuth: []
      summary: Delete a camera
      tags:
      - camera
  /cameras:
    get:
      consumes:
      - application/json
      description: Retrieve cameras with optional query parameters for filtering and
        sorting
      parameters:
      - description: Sort order
        enum:
        - alphabetical
        - counts
        in: query
        name: sort
        type: string
      - description: Number of results to return
        in: query
        name: page_size
        type: integer
      - description: Filter by camera make
        in: query
        name: make
        type: string
      - description: Filter by camera model
        in: query
        name: model
        type: string
      - description: Filter by specific camera ID
        in: query
        name: id
        type: integer
      - description: Include count data
        in: query
        name: include_counts
        type: boolean
      - description: Exclude zero counts
        in: query
        name: exclude_zero_counts
        type: boolean
      - description: Only return entries with at least this many posts, implies include_counts
        in: query
        name: min_count
        type: integer
      - description: Attach this many top scoring posts to each entry (1 to 10)
        in: query
        name: top_posts
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CamerasResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/analogdb.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/analogdb.Error'
      summary: Get cameras with optional filtering
      tags:
      - cameras
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"